// es/config.d.ts
declare const CONFIG: {
    appId: string;
    prefix: string;
    mtrVer: string;
    stat_share_app: boolean;
    stat_pull_down_fresh: boolean;
    stat_page_scroll: boolean;
    stat_hide: boolean;
    stat_unload: boolean;
    stat_reach_bottom: boolean;
    stat_auto_click: boolean;
    stat_location: boolean;
    stat_api: boolean;
    stat_api_success: boolean;
    stat_api_long_time: number;
    stat_api_fail: boolean;
    scene: {
        "1000": string;
        "1002": string;
        "1005": string;
        "1007": string;
        "1011": string;
        "1014": string;
        "1020": string;
        "1023": string;
        "1037": string;
        "1038": string;
        "1090": string;
        "1200": string;
        "1201": string;
        "1202": string;
        "1203": string;
        "1204": string;
        "1205": string;
        "1206": string;
        "1207": string;
        "1208": string;
        "1209": string;
        "1300": string;
        "0000": string;
    };
};
export default CONFIG;